+++ custom_components/heru/const.py
@@ -392,48 +392,6 @@ HERU_SENSORS = [
         "entity_category": None,
         "register_type": INPUT_REGISTERS_BINARY,
     },
-    {
-        "name": "Quality sensor 1 type",
-        "modbus_address": "3x00041",
-        "scale": None,
-        "icon": ICON_AIR_PURIFIER,
-        "unit_of_measurement": None,
-        "entity_category": None,
-        "device_class": SensorDeviceClass.ENUM,
-        "state_class": None,
-        "entity_category": EntityCategory.DIAGNOSTIC,
-        "register_type": INPUT_REGISTERS,
-        "options": ["None", "RH", "CO2", "VOC"],
-        "entity_registry_enabled_default": False,
-    },
-    {
-        "name": "Quality sensor 2 type",
-        "modbus_address": "3x00043",
-        "scale": None,
-        "icon": ICON_AIR_PURIFIER,
-        "unit_of_measurement": None,
-        "entity_category": None,
-        "device_class": SensorDeviceClass.ENUM,
-        "state_class": None,
-        "entity_category": EntityCategory.DIAGNOSTIC,
-        "register_type": INPUT_REGISTERS,
-        "options": ["None", "RH", "CO2", "VOC"],
-        "entity_registry_enabled_default": False,
-    },
-    {
-        "name": "Quality sensor 3 type",
-        "modbus_address": "3x00045",
-        "scale": None,
-        "icon": ICON_AIR_PURIFIER,
-        "unit_of_measurement": None,
-        "entity_category": None,
-        "device_class": SensorDeviceClass.ENUM,
-        "state_class": None,
-        "entity_category": EntityCategory.DIAGNOSTIC,
-        "register_type": INPUT_REGISTERS,
-        "options": ["None", "RH", "CO2", "VOC"],
-        "entity_registry_enabled_default": False,
-    },
 ]
